+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Changes
 
 * Add Elasticsearch bucket script pipeline aggregator
 * Added ability to hide metrics for Elasticsearch MetricAggs
+* Add derivative metric aggregation for Elasticsearch
 * ...